Skip exisisting features during bert gen

+++ bert_gen.py
@@ -1,3 +1,4 @@
+import os
 import torch
 from multiprocessing import Pool
 import commons
@@ -36,15 +37,16 @@ def process_line(x):
         word2ph[0] += 1
 
     bert_path = wav_path.replace(".WAV", ".wav").replace(".wav", ".bert.pt")
-
-    try:
-        bert = torch.load(bert_path)
-        assert bert.shape[0] == 2048
-    except Exception:
-        bert = get_bert(text, word2ph, language_str, device)
-        assert bert.shape[-1] == len(phone)
-        torch.save(bert, bert_path)
-
+    if os.path.isfile(bert_path):
+        try:
+            bert = torch.load(bert_path)
+            assert bert.shape[0] == 2048
+        except Exception:
+            bert = get_bert(text, word2ph, language_str, device)
+            assert bert.shape[-1] == len(phone)
+            torch.save(bert, bert_path)
+    else:
+        print(f"{bert_path} already exsists.")
 
 preprocess_text_config = config.preprocess_text_config
